Convert ImageUploader to TypeScript via the shared icon control
Replace ImageUploader.js with a .tsx wrapper that uses the
createIconControl helper, so the bar control logic stays in lockstep
with the four adder controls. Widen ImageUploaderDialog.missionId to
number | string to match the rest of the codebase.
